+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - unify usage of v-models and emits in codebase [#305](https://github.com/upb-uc4/ui-web/pull/305)
 - increase performance of course loading [#317](https://github.com/upb-uc4/ui-web/pull/317)
 - move course filtering for lecturer from frontend to backend [#317](https://github.com/upb-uc4/ui-web/pull/317)
+- remove unused wrapper components for asynchronous data loading [#341](https://github.com/upb-uc4/ui-web/pull/341)
   
 ## Bugfix
 - Fix a possible XSS vulnerability [#141](https://github.com/upb-uc4/ui-web/issues/141)